@@ -44,18 +44,18 @@
   },
   "devDependencies": {
     "@cowtech/eslint-config": "^9.0.3",
-    "@swc/cli": "^0.3.9",
-    "@swc/core": "^1.4.0",
+    "@swc/cli": "^0.3.12",
+    "@swc/core": "^1.4.13",
     "@types/http-errors": "^2.0.4",
-    "@types/node": "^20.11.16",
+    "@types/node": "^20.12.7",
     "c8": "^9.1.0",
     "chokidar": "^3.6.0",
     "concurrently": "^8.2.2",
     "eslint": "^8.57.0",
     "http-errors": "^2.0.0",
     "prettier": "^3.2.5",
-    "tsx": "^4.7.0",
-    "typescript": "^5.3.3"
+    "tsx": "^4.7.2",
+    "typescript": "^5.4.5"
   },
   "engines": {
     "node": ">= 18.18.0"
