@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   printWidth: 120,
   semi: false,
   singleQuote: true,
